+++ test/fakeListenerInvocation.ts
@@ -14,16 +14,17 @@
  * limitations under the License.
  */
 
-import { RemoteRepoRef } from "@atomist/automation-client/operations/common/RepoId";
-import { GitProject } from "@atomist/automation-client/project/git/GitProject";
-import { Project } from "@atomist/automation-client/project/Project";
+import {
+    GitProject,
+    Project,
+} from "@atomist/automation-client";
 import { PushListenerInvocation } from "@atomist/sdm";
 
 export function fakeListenerInvocation(project: Project): PushListenerInvocation {
     return {
-        push: {id: new Date().getTime() + "_", branch: "master"},
+        push: {id: Date.now() + "_", branch: "master"},
         project: project as any as GitProject,
-        id: project.id as RemoteRepoRef,
+        id: project.id as any,
         context: null,
         addressChannels: null,
         credentials: null,