+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 -   Consider reading defaults from metadata rather than hardcoding them. [#40](https://github.com/atomist/sdm-pack-spring/issues/40)
 -   Move from src to lib in dependencies. [#48](https://github.com/atomist/sdm-pack-spring/issues/48)
 
+### Fixed
+
+-   Use cross-spawn. [#51](https://github.com/atomist/sdm-pack-spring/issues/51)
+
 ## [1.0.0-M.3](https://github.com/atomist/sdm-pack-spring/compare/1.0.0-M.1...1.0.0-M.3) - 2018-09-04
 
 ### Added